dirty migration: modify balance and amount types

+++ schema.sql
@@ -1,7 +1,7 @@
 -- Create users table
 CREATE TABLE users (
     id BIGINT PRIMARY KEY,
-    balance DECIMAL(15, 2) NOT NULL DEFAULT 0.00 CHECK ( balance >= 0.00 ),
+    balance BIGINT NOT NULL DEFAULT 0 CHECK ( balance >= 0 ),
     created_at TIMESTAMP NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP NOT NULL DEFAULT NOW()
 ); 
@@ -13,7 +13,7 @@ CREATE TABLE transactions (
     transaction_id VARCHAR(255) UNIQUE NOT NULL,
     source_type VARCHAR(50) NOT NULL,
     state VARCHAR(10) NOT NULL,
-    amount DECIMAL(15, 2) NOT NULL,
+    amount BIGINT NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT NOW(),
     FOREIGN KEY (user_id) REFERENCES users(id)
 );
